shader no. 465

+++ src/shaders/usefulFunctions/usefulFunctions.js
@@ -270,16 +270,16 @@ float sdEgg(vec2 p, float ra, float rb)
             length(vec2(p.x + r, p.y)) -2.0 * r) - rb;
 }
 
-float sdPolygon(vec2 p, float sides, float scale)
+float sdPolygon(vec2 p, int sides, float scale)
 {
     p = p * 2. - 1.;
     float angle = atan(p.x, p.y) + PI;
-    float radius = TWO_PI/sides;
+    float radius = TWO_PI/float(sides);
     float d = cos(floor(.5 + angle/ radius) * radius - angle) * length(p);
     return 1. - smoothstep(scale, scale + 0.01, d); 
 }
 
-float sdPolygonOutline(vec2 p, float sides, float scale)
+float sdPolygonOutline(vec2 p, int sides, float scale)
 {
     float x = sdPolygon(p, sides, scale);
     float y = sdPolygon(p, sides, scale + 0.05);
